Remove unnecessary static cast in 'main' source file

+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include "chip8.hpp"
 
 int main(int argc, char *argv[]) {
-    return static_cast<int>(chip8::run(argc, argv));
+    return chip8::run(argc, argv);
 }
